Update chairController
de-bugged and working. hold a trigger closed for an amount of time described by door_open_delay and after the delay ( which can be abandoned by releasing the trigger which resets ) two relays turn on and a third turns on and off to "puppeteer" a door open. that relay code is written with blocking type delay( ); codes and cant be interrupted. the initial two relays remain on until the trigger is releases and then the door is shut by a fourth relay that also uses delay( ); type code to close the door uninterrupted. then it resets. additionally the door delay can be bypassed when the trigger is closed by a signal on pin 3 which will be a handheld remote. i might change it to skip to door_open from any point in the program. the door delay will be set much longer like 30-40 seconds and is a test of patience . haha